codex_pod_agent/requirements.txt
anthropic>=0.32.0
feedparser>=6.0.11
openai>=1.40.0
pydub>=0.25.1
edge-tts>=6.1.14
requests>=2.32.0
beautifulsoup4>=4.12.3
readability-lxml>=0.8.1
lxml>=4.9.4
PyYAML>=6.0.2
